api/bot_autentification.py
import os
import pickle
from googleapiclient.discovery import build
from google_auth_oauthlib.flow import InstalledAppFlow
from google.auth.transport.requests import Request
from api.bot_utils import *



# Función para autenticar con Google Drive
def authenticate_google(user_id):
    """Autentica al usuario con Google Drive y retorna el cliente de la API."""
    creds = None
    token_path = f'token_{user_id}.pickle'

    if os.path.exists(token_path):
        with open(token_path, 'rb') as token:
            creds = pickle.load(token)

    if not creds or not creds.valid:
        if creds and creds.expired and creds.refresh_token:
            creds.refresh(Request())
        else:
            flow = InstalledAppFlow.from_client_secrets_file(CREDENTIALS_FILE, SCOPES)
            creds = flow.run_console()  # Usar run_console para permitir autorización en dispositivo

        with open(token_path, 'wb') as token:
            pickle.dump(creds, token)

    return build('drive', 'v3', credentials=creds)

